requirements.txt
numpy
click
pillow
#torch
pandas
pytorch-lightning
transformers